finnhub-producer/requirements.txt
websocket-client==1.6.1
kafka-python==2.0.2
